+++ cmd/micro-cli/cli.go
@@ -1,4 +1,4 @@
-package cli
+package microcli
 
 import (
 	"bufio"
@@ -6,8 +6,11 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"sort"
+	"os/exec"
+	"os/signal"
+	"path/filepath"
 	"strings"
+	"syscall"
 
 	"github.com/urfave/cli/v2"
 	"go-micro.dev/v5/client"
@@ -15,7 +18,6 @@ import (
 	"go-micro.dev/v5/codec/bytes"
 	"go-micro.dev/v5/registry"
 
-	"github.com/micro/micro/v5/cmd/micro-cli/new"
 	"github.com/micro/micro/v5/cmd/micro-cli/util"
 )
 
@@ -25,147 +27,285 @@ var (
 	version = "5.0.0-dev"
 )
 
-type Command struct {
-	Name   string
-	Usage  string
-	Action func(*cli.Context, []string) error
+func genProtoHandler(c *cli.Context) error {
+	cmd := exec.Command("find", ".", "-name", "*.proto", "-exec", "protoc", "--proto_path=.", "--micro_out=.", "--go_out=.", `{}`, `;`)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	return cmd.Run()
 }
 
-var commands = []Command{
-	{
-		Name:  "services",
-		Usage: "List available services",
-		Action: func(ctx *cli.Context, args []string) error {
-			services, err := registry.ListServices()
-			if err != nil {
-				return err
-			}
-			for _, service := range services {
-				fmt.Println(service.Name)
-			}
-			return nil
-		},
-	},
-	{
-		Name:  "call",
-		Usage: "Call a service",
-		Action: func(ctx *cli.Context, args []string) error {
-			if len(args) < 2 {
-				return fmt.Errorf("Usage: [service] [endpoint] [request]")
-			}
-
-			service := args[0]
-			endpoint := args[1]
-			request := `{}`
-
-			// get the request if present
-			if len(args) >= 3 {
-				request = strings.Join(args[2:], " ")
-			}
-
-			req := client.NewRequest(service, endpoint, &bytes.Frame{Data: []byte(request)})
-			var rsp bytes.Frame
-			err := client.Call(context.TODO(), req, &rsp)
-			if err != nil {
-				return err
-			}
-
-			fmt.Print(string(rsp.Data))
-			return nil
-		},
-	},
-	{
-		Name:  "describe",
-		Usage: "Describe a service",
-		Action: func(ctx *cli.Context, args []string) error {
-			if len(args) != 1 {
-				return fmt.Errorf("Usage: [service]")
-			}
-
-			service := args[0]
-			services, err := registry.GetService(service)
-			if err != nil {
-				return err
-			}
-			if len(services) == 0 {
-				return nil
-			}
-			b, _ := json.MarshalIndent(services[0], "", "    ")
-			fmt.Println(string(b))
-			return nil
-		},
-	},
-}
-
-func Run(c *cli.Context) error {
-	reader := bufio.NewReader(os.Stdin)
-
-	commandMap := map[string]Command{}
-	helpUsage := []string{}
-
-	for _, c := range commands {
-		commandMap[c.Name] = c
-		helpUsage = append(helpUsage, fmt.Sprintf("%-20s%s", c.Name, c.Usage))
+func lastNonEmptyLine(s string) string {
+	lines := strings.Split(s, "\n")
+	for i := len(lines) - 1; i >= 0; i-- {
+		if strings.TrimSpace(lines[i]) != "" {
+			return lines[i]
+		}
 	}
+	return ""
+}
 
-	sort.Strings(helpUsage)
-
-	for {
-		fmt.Print("micro> ") // Print the prompt
-
-		input, _ := reader.ReadString('\n') // Read input until a newline
-		input = input[:len(input)-1]        // Remove the trailing newline
-
-		args := strings.Split(input, " ")
-
-		if len(args) == 0 {
-			continue
+func lastLogLine(path string) string {
+	f, err := os.Open(path)
+	if err != nil {
+		return ""
+	}
+	defer f.Close()
+	var last string
+	scan := bufio.NewScanner(f)
+	for scan.Scan() {
+		if strings.TrimSpace(scan.Text()) != "" {
+			last = scan.Text()
 		}
+	}
+	return last
+}
 
-		command := args[0]
-
-		if command == "exit" {
-			fmt.Println("Exiting...")
-			return nil
-		}
-
-		if v, ok := commandMap[command]; ok {
-			err := v.Action(c, args[1:])
-			if err != nil {
-				fmt.Println(err)
+func waitAndCleanup(procs []*exec.Cmd, pidFiles []string) {
+	ch := make(chan os.Signal, 1)
+	signal.Notify(ch, os.Interrupt)
+	go func() {
+		<-ch
+		for _, proc := range procs {
+			if proc.Process != nil {
+				_ = proc.Process.Kill()
 			}
-			continue
 		}
-
-		if command == "help" || command == "?" {
-			fmt.Println("Commands:")
-			fmt.Println(strings.Join(helpUsage, "\n"))
-			continue
+		for _, pf := range pidFiles {
+			_ = os.Remove(pf)
 		}
-
-		if srv, err := util.LookupService(command); err != nil {
-			fmt.Println(util.CliError(err))
-		} else if srv != nil && util.ShouldRenderHelp(args) {
-			fmt.Println(cli.Exit(util.FormatServiceUsage(srv, c), 0))
-		} else if srv != nil {
-			err := util.CallService(srv, args)
-			if err != nil {
-				fmt.Println(util.CliError(err))
-			}
+		os.Exit(1)
+	}()
+	for i, proc := range procs {
+		_ = proc.Wait()
+		if proc.Process != nil {
+			_ = os.Remove(pidFiles[i])
 		}
 	}
 }
 
 func init() {
-	cmd.Register(&cli.Command{
-		Name:   "cli",
-		Usage:  "Launch the interactive CLI",
-		Action: Run,
-	})
-	cmd.Register(&cli.Command{
-		Name:        "new",
-		Usage:       "Create a new service",
-		Description: `'micro new' generates a new service skeleton. Example: 'micro new helloworld && cd helloworld'`,
-		Action:      new.Run,
-	})
+	cmd.Register([]*cli.Command{
+		{
+			Name:  "gen",
+			Usage: "Generate various things",
+			Subcommands: []*cli.Command{
+				{
+					Name:   "proto",
+					Usage:  "Generate proto requires protoc and protoc-gen-micro",
+					Action: genProtoHandler,
+				},
+			},
+		},
+		{
+			Name:  "services",
+			Usage: "List available services",
+			Action: func(ctx *cli.Context) error {
+				services, err := registry.ListServices()
+				if err != nil {
+					return err
+				}
+				for _, service := range services {
+					fmt.Println(service.Name)
+				}
+				return nil
+			},
+		},
+		{
+			Name:  "call",
+			Usage: "Call a service",
+			Action: func(ctx *cli.Context) error {
+				args := ctx.Args()
+
+				if args.Len() < 2 {
+					return fmt.Errorf("Usage: [service] [endpoint] [request]")
+				}
+
+				service := args.Get(0)
+				endpoint := args.Get(1)
+				request := `{}`
+
+				if args.Len() == 3 {
+					request = args.Get(2)
+				}
+
+				req := client.NewRequest(service, endpoint, &bytes.Frame{Data: []byte(request)})
+				var rsp bytes.Frame
+				err := client.Call(context.TODO(), req, &rsp)
+				if err != nil {
+					return err
+				}
+
+				fmt.Print(string(rsp.Data))
+				return nil
+			},
+		},
+		{
+			Name:  "describe",
+			Usage: "Describe a service",
+			Action: func(ctx *cli.Context) error {
+				args := ctx.Args()
+
+				if args.Len() != 1 {
+					return fmt.Errorf("Usage: [service]")
+				}
+
+				service := args.Get(0)
+				services, err := registry.GetService(service)
+				if err != nil {
+					return err
+				}
+				if len(services) == 0 {
+					return nil
+				}
+				b, _ := json.MarshalIndent(services[0], "", "    ")
+				fmt.Println(string(b))
+				return nil
+			},
+		},
+		{
+			Name:  "status",
+			Usage: "Check status of running services",
+			Action: func(ctx *cli.Context) error {
+				homeDir, err := os.UserHomeDir()
+				if err != nil {
+					return fmt.Errorf("failed to get home dir: %w", err)
+				}
+				runDir := filepath.Join(homeDir, "micro", "run")
+				files, err := os.ReadDir(runDir)
+				if err != nil {
+					return fmt.Errorf("failed to read run dir: %w", err)
+				}
+				fmt.Printf("%-20s %-8s %-8s %s\n", "SERVICE", "PID", "STATUS", "DIRECTORY")
+				for _, f := range files {
+					if f.IsDir() || !strings.HasSuffix(f.Name(), ".pid") {
+						continue
+					}
+					service := f.Name()[:len(f.Name())-4]
+					pidFilePath := filepath.Join(runDir, f.Name())
+					pidFile, err := os.Open(pidFilePath)
+					if err != nil {
+						continue
+					}
+					var pid int
+					var dir, reason string
+					fmt.Fscanf(pidFile, "%d\n%s\nreason: [%s]\n", &pid, &dir, &reason)
+					pidFile.Close()
+					status := "stopped"
+					if pid > 0 {
+						proc, err := os.FindProcess(pid)
+						if err == nil {
+							if err := proc.Signal(syscall.Signal(0)); err == nil {
+								status = "running"
+							}
+						}
+					}
+					if reason != "" && status != "running" {
+						fmt.Printf("%-20s %-8d %-8s %-40s %s\n", service, pid, status, reason, dir)
+					} else {
+						fmt.Printf("%-20s %-8d %-8s %-40s %s\n", service, pid, status, "", dir)
+					}
+				}
+				return nil
+			},
+		},
+		{
+			Name:  "stop",
+			Usage: "Stop a running service",
+			Action: func(ctx *cli.Context) error {
+				if ctx.Args().Len() != 1 {
+					return fmt.Errorf("Usage: micro stop [service]")
+				}
+				service := ctx.Args().Get(0)
+				homeDir, err := os.UserHomeDir()
+				if err != nil {
+					return fmt.Errorf("failed to get home dir: %w", err)
+				}
+				runDir := filepath.Join(homeDir, "micro", "run")
+				pidFilePath := filepath.Join(runDir, service+".pid")
+				pidFile, err := os.Open(pidFilePath)
+				if err != nil {
+					return fmt.Errorf("no pid file for service %s", service)
+				}
+				var pid int
+				var dir, reason string
+				fmt.Fscanf(pidFile, "%d\n%s\nreason: [%s]\n", &pid, &dir, &reason)
+				pidFile.Close()
+				if pid <= 0 {
+					_ = os.Remove(pidFilePath)
+					return fmt.Errorf("service %s is not running", service)
+				}
+				proc, err := os.FindProcess(pid)
+				if err != nil {
+					_ = os.Remove(pidFilePath)
+					return fmt.Errorf("could not find process for %s", service)
+				}
+				if err := proc.Signal(syscall.SIGTERM); err != nil {
+					_ = os.Remove(pidFilePath)
+					return fmt.Errorf("failed to stop service %s: %v", service, err)
+				}
+				_ = os.Remove(pidFilePath)
+				fmt.Printf("Stopped service %s (pid %d)\n", service, pid)
+				return nil
+			},
+		},
+		{
+			Name:  "logs",
+			Usage: "Show logs for a service",
+			Action: func(ctx *cli.Context) error {
+				if ctx.Args().Len() != 1 {
+					return fmt.Errorf("Usage: micro logs [service]")
+				}
+				service := ctx.Args().Get(0)
+				homeDir, err := os.UserHomeDir()
+				if err != nil {
+					return fmt.Errorf("failed to get home dir: %w", err)
+				}
+				logsDir := filepath.Join(homeDir, "micro", "logs")
+				logFilePath := filepath.Join(logsDir, service+".log")
+				f, err := os.Open(logFilePath)
+				if err != nil {
+					return fmt.Errorf("could not open log file for service %s: %v", service, err)
+				}
+				defer f.Close()
+				scan := bufio.NewScanner(f)
+				for scan.Scan() {
+					fmt.Println(scan.Text())
+				}
+				return scan.Err()
+			},
+		},
+	}...)
+
+	cmd.App().Action = func(c *cli.Context) error {
+		if c.Args().Len() == 0 {
+			return nil
+		}
+
+		v, err := exec.LookPath("micro-" + c.Args().First())
+		if err == nil {
+			ce := exec.Command(v, c.Args().Slice()[1:]...)
+			ce.Stdout = os.Stdout
+			ce.Stderr = os.Stderr
+			return ce.Run()
+		}
+
+		command := c.Args().Get(0)
+		args := c.Args().Slice()
+
+		if srv, err := util.LookupService(command); err != nil {
+			return util.CliError(err)
+		} else if srv != nil && util.ShouldRenderHelp(args) {
+			return cli.Exit(util.FormatServiceUsage(srv, c), 0)
+		} else if srv != nil {
+			err := util.CallService(srv, args)
+			return util.CliError(err)
+		}
+
+		return nil
+	}
+
+	cmd.Init(
+		cmd.Name("micro"),
+		cmd.Version(version),
+	)
 }
